requirements.txt
asgiref==3.5.2
Django==4.0.6
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
unicodecsv==0.14.1
